requirements.txt
requests~=2.31.0
pyModbusTCP~=0.2.0
wallbox~=0.8.0
flask~=3.0.0
flask-restx~=1.3.0
influxdb-client~=1.36.0
flake8~=7.0.0
questionary~=2.0.1  # Used in configure.py
pyyaml~=6.0.1      # Used for YAML config files
python-dateutil~=2.8.2  # Commonly needed for date handling
typing-extensions>=4.12.2  # Changed to match pydantic's requirement
pytest~=7.4.4  # For running tests
